+++ clang/lib/Sema/SemaDeclCXX.cpp
@@ -11323,8 +11323,16 @@ Decl *Sema::ActOnConversionDeclarator(CXXConversionDecl *Conversion) {
   }
 
   if (FunctionTemplateDecl *ConversionTemplate
-                                = Conversion->getDescribedFunctionTemplate())
+                                = Conversion->getDescribedFunctionTemplate()) {
+    if (ConvType->isUndeducedAutoType()) {
+      Diag(Conversion->getTypeSpecStartLoc(), diag::err_auto_not_allowed)
+        << Conversion->getTypeSourceInfo()->getTypeLoc().getContainedAutoTypeLoc().getSourceRange()
+        << llvm::to_underlying(Conversion->getConversionType()->getAs<AutoType>()->getKeyword())
+        << /* in declaration of conversion function template */ 24;
+    }
+
     return ConversionTemplate;
+  }
 
   return Conversion;
 }